MCOL-174 fix compile errors

+++ utils/funcexp/func_mod.cpp
@@ -75,7 +75,7 @@ IDB_Decimal Func_mod::getDecimalVal(Row& row,
 
     IDB_Decimal d = parm[0]->data()->getDecimalVal(row, isNull);
     int64_t value = d.value / pow(10.0, d.scale);
-    int lefto = d.value % pow(10.0, d.scale);
+    int lefto = d.value % (int)pow(10.0, d.scale);
 
     int64_t mod = (value % div) * pow(10.0, d.scale) + lefto;
